Create full Axaml/Axaml.cs controls for ComboBox/ProgressRing/RadioButton
